Block mobile access to wifi
The script searches in the list of dhcp-server leases all the connected clients with the flag "dynamic=yes" and where the dhcp server is the one defined in the DHCPSERVER variable. This is in case we have more than 1 dhcp server in order not to filter all of them. Look for the host-name matches "Android", "iPad", etc, then add a new lease with the flag "block-access=yes" for that mac. Finally release all leases with that name to not occupy an IP address.